address pr comments, fixing tutorial
