+++ packages/rmf-dashboard-framework/package.json
@@ -91,5 +91,9 @@
     "typescript-json-schema": "^0.64.0",
     "vite": "^5.3.5",
     "vitest": "^2.0.4"
+  },
+  "exports": {
+    "./components": "./src/components",
+    "./micro-apps": "./src/micro-apps"
   }
 }
