+++ packages/rmf-dashboard-framework/src/components/index.ts
@@ -1,1 +1,3 @@
+export * from './micro-app';
 export * from './rmf-dashboard';
+export * from './workspace';
