+++ packages/rmf-dashboard-framework/src/micro-apps/map-app.ts
@@ -1,5 +1,5 @@
+import { createMicroApp, MicroAppManifest } from '../components';
 import type { MapProps } from '../components/map';
-import { createMicroApp, MicroAppManifest } from '.';
 
 export default function createMapApp(config: MapProps): MicroAppManifest {
   return createMicroApp(
